Split internal/external schema tests
