+++ internal/schema/schema_test.go
@@ -1,110 +1,87 @@
-package schema
+package schema_test
 
 import (
-	"strings"
 	"testing"
-	"text/scanner"
 
-	"github.com/graph-gophers/graphql-go/errors"
-	"github.com/graph-gophers/graphql-go/internal/common"
+	"github.com/graph-gophers/graphql-go/internal/schema"
 )
 
-// TestParseObjectDef tests the logic for parsing object types from the schema definition as
-// written in `parseObjectDef()`.
-func TestParseObjectDef(t *testing.T) {
-	type testCase struct {
-		description string
-		definition  string
-		expected    *Object
-		err         *errors.QueryError
+type parseTestCase struct {
+	description string
+	sdl         string
+	expected    *schema.Schema
+	err         error
+}
+
+var parseTests = []parseTestCase{{
+	description: "Parses type with description string",
+	sdl: `
+	"Single line description."
+	type Type {
+		field: String
+	}`,
+	expected: &schema.Schema{
+		Types: map[string]schema.NamedType{
+			"Type": &schema.Object{
+				Name: "Type",
+				Desc: "Single line description.",
+			},
+		},
+	},
+}, {
+	description: "Parses type with multi-line description string",
+	sdl: `
+	"""
+	Multi-line description.
+	"""
+	type Type {
+		field: String
+	}`,
+	expected: &schema.Schema{
+		Types: map[string]schema.NamedType{
+			"Type": &schema.Object{
+				Name: "Type",
+				Desc: "Multi-line description.",
+			},
+		},
+	},
+}, {
+	description: "Parses type with multi-line description and ignores comments",
+	sdl: `
+	"""
+	Multi-line description with ignored comments.
+	"""
+	# This comment should be ignored.
+	type Type {
+		field: String
+	}`,
+	expected: &schema.Schema{
+		Types: map[string]schema.NamedType{
+			"Type": &schema.Object{
+				Name: "Type",
+				Desc: "Multi-line description with ignored comments.",
+			},
+		},
+	},
+}}
+
+func TestParse(t *testing.T) {
+	setup := func(t *testing.T) *schema.Schema {
+		t.Helper()
+		return schema.New()
 	}
 
-	tests := []testCase{{
-		description: "Parses type inheriting single interface",
-		definition:  "Hello implements World { field: String }",
-		expected:    &Object{Name: "Hello", interfaceNames: []string{"World"}},
-	}, {
-		description: "Parses type inheriting multiple interfaces",
-		definition:  "Hello implements Wo & rld { field: String }",
-		expected:    &Object{Name: "Hello", interfaceNames: []string{"Wo", "rld"}},
-	}, {
-		description: "Parses type inheriting multiple interfaces with leading ampersand",
-		definition:  "Hello implements & Wo & rld { field: String }",
-		expected:    &Object{Name: "Hello", interfaceNames: []string{"Wo", "rld"}},
-	}, {
-		description: "Allows legacy SDL interfaces",
-		definition:  "Hello implements Wo, rld { field: String }",
-		expected:    &Object{Name: "Hello", interfaceNames: []string{"Wo", "rld"}},
-	}}
-
-	setup := func(def string) *common.Lexer {
-		sc := &scanner.Scanner{
-			Mode: scanner.ScanIdents | scanner.ScanInts | scanner.ScanFloats | scanner.ScanStrings,
-		}
-		sc.Init(strings.NewReader(def))
-		return common.NewLexer(sc)
-	}
-
-	for _, test := range tests {
+	for _, test := range parseTests {
 		t.Run(test.description, func(t *testing.T) {
-			var actual *Object
-			lex := setup(test.definition)
+			t.Skip("TODO: add support for descriptions")
+			schema := setup(t)
 
-			parse := func() { actual = parseObjectDef(lex) }
-			err := lex.CatchSyntaxError(parse)
+			err := schema.Parse(test.sdl)
+			if err != nil {
+				t.Fatal(err)
+			}
 
-			compareErrors(t, test.err, err)
-			compareObjects(t, test.expected, actual)
+			// TODO: verify schema is the same as expected.
 		})
-	}
-}
-
-func compareErrors(t *testing.T, expected, actual *errors.QueryError) {
-	t.Helper()
-
-	switch {
-	case expected != nil && actual != nil:
-		if expected.Message != actual.Message {
-			t.Fatalf("wanted error message %q, got %q", expected.Message, actual.Message)
-		}
-		// TODO: Check error locations are as expected.
-
-	case expected != nil && actual == nil:
-		t.Fatalf("missing expected error: %q", expected)
-
-	case expected == nil && actual != nil:
-		t.Fatalf("got unexpected error: %q", actual)
-	}
-}
-
-func compareObjects(t *testing.T, expected, actual *Object) {
-	t.Helper()
-
-	switch {
-	case expected == nil && expected == actual:
-		return
-	case expected == nil && actual != nil:
-		t.Fatalf("got an unexpected object: %#v", actual)
-	case expected != nil && actual == nil:
-		t.Fatalf("wanted non-nil object, got nil")
-	}
-
-	if expected.Name != actual.Name {
-		t.Errorf("wrong object name: want %q, got %q", expected.Name, actual.Name)
-	}
-
-	if len(expected.interfaceNames) != len(actual.interfaceNames) {
-		t.Fatalf(
-			"wrong number of interface names: want %s, got %s",
-			expected.interfaceNames,
-			actual.interfaceNames,
-		)
-	}
-
-	for i, expectedName := range expected.interfaceNames {
-		actualName := actual.interfaceNames[i]
-		if expectedName != actualName {
-			t.Errorf("wrong interface name: want %q, got %q", expectedName, actualName)
-		}
 	}
 }
